Add files via upload
Sunspots are temporary phenomena on the Sun's photosphere that appear as spots darker than the 
surrounding areas. They are regions of reduced surface temperature caused by concentrations of 
magnetic field flux that inhibit convection.
Tested the forecasting model in three (3) distinct datasets. On Daily, Monthly Mean, and Yearly 
Mean sunspots.
Since the time unit (day, month, year) varies from dataset to dataset, made the code agnostic of 
the input. In other words, the code determines the unit of the time series.
2) Then, trained the model (on the respective dataset) and predicted the Sunspots values from the last 
date of the dataset into X units of time into the future.
a) Should the unit of time be day, then predict the # of sunspots for 100/200/365 days into the 
future.
b) Should the unit of time be month, then predict the # of sunspots for 1/6/9 months into the 
future.
c) Should the unit of time be year, then predict the # of sunspots for 1/10/20 years into the 
future.
3) Tuned FBProphet model on the following parameters:
a) Forecasting growth: Plausible values = logistic; linear; flat